feat - add calc expense and save mistake record

+++ database/migrations/2021_11_22_144654_create_invest_futures_table.php
@@ -43,7 +43,7 @@ class CreateInvestFuturesTable extends Migration
             ->default(0)
             ->comment('最終損益 min(權益損益, 沖銷損益)');
 
-        $table->smallInteger('total_quota')
+        $table->decimal('total_quota', 5, 1)
             ->unsigned()
             ->default(0)
             ->comment('總分額數');
